+++ apps/posts/src/views/PostAnalytics/Growth/components/growth-sources.tsx
@@ -1,6 +1,7 @@
+import DisabledSourcesIndicator from '../../components/disabled-sources-indicator';
 import React from 'react';
 import SourceIcon from '../../components/source-icon';
-import {BaseSourceData, ProcessedSourceData, extendSourcesWithPercentages, processSources, useNavigate} from '@tryghost/admin-x-framework';
+import {BaseSourceData, ProcessedSourceData, extendSourcesWithPercentages, processSources} from '@tryghost/admin-x-framework';
 import {Button, Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle, EmptyIndicator, LucideIcon, Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle, SheetTrigger, Table, TableBody, TableCell, TableHead, TableHeader, TableRow, cn, formatNumber} from '@tryghost/shade';
 import {useAppContext} from '@src/providers/posts-app-context';
 
@@ -106,7 +107,6 @@ export const GrowthSources: React.FC<SourcesCardProps> = ({
     className
 }) => {
     const {appSettings} = useAppContext();
-    const navigate = useNavigate();
     // Process and group sources data with pre-computed icons and display values
     const processedData = React.useMemo(() => {
         return processSources({
@@ -151,18 +151,7 @@ export const GrowthSources: React.FC<SourcesCardProps> = ({
             }
             <CardContent>
                 {mode === 'growth' && !appSettings?.analytics.membersTrackSources ? (
-                    <EmptyIndicator
-                        actions={
-                            <Button variant='outline' onClick={() => navigate('/settings/analytics', {crossApp: true})}>
-                                Open settings
-                            </Button>
-                        }
-                        className='py-10'
-                        description='Enable member source tracking in settings to see which content drives member growth.'
-                        title='Member sources have been disabled'
-                    >
-                        <LucideIcon.Activity />
-                    </EmptyIndicator>
+                    <DisabledSourcesIndicator className='py-10' />
                 ) : topSources.length > 0 ? (
                     <SourcesTable
                         data={topSources}
