+++ apps/stats/src/views/Stats/Growth/components/growth-sources.tsx
@@ -1,8 +1,9 @@
+import DisabledSourcesIndicator from '../../components/disabled-sources-indicator';
 import React, {useState} from 'react';
 import SortButton from '../../components/sort-button';
 import SourceIcon from '../../components/source-icon';
 import {Button, EmptyIndicator, LucideIcon, Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle, SheetTrigger, Skeleton, Table, TableBody, TableCell, TableFooter, TableHead, TableHeader, TableRow, centsToDollars, formatNumber} from '@tryghost/shade';
-import {getFaviconDomain, getSymbol, useAppContext, useNavigate} from '@tryghost/admin-x-framework';
+import {getFaviconDomain, getSymbol, useAppContext} from '@tryghost/admin-x-framework';
 import {getPeriodText} from '@src/utils/chart-helpers';
 import {useGlobalData} from '@src/providers/global-data-provider';
 import {useMrrHistory} from '@tryghost/admin-x-framework/api/stats';
@@ -94,7 +95,6 @@ export const GrowthSources: React.FC<GrowthSourcesProps> = ({
     const {data: globalData} = useGlobalData();
     const {data: mrrHistoryResponse} = useMrrHistory();
     const {appSettings} = useAppContext();
-    const navigate = useNavigate();
 
     // Use external sort state if provided, otherwise use internal state
     const [internalSortBy, setInternalSortBy] = useState<SourcesOrder>('free_members desc');
@@ -169,17 +169,7 @@ export const GrowthSources: React.FC<GrowthSourcesProps> = ({
             <TableBody>
                 <TableRow className='last:border-none'>
                     <TableCell className='border-none py-12 group-hover:!bg-transparent' colSpan={appSettings?.paidMembersEnabled ? 4 : 2}>
-                        <EmptyIndicator
-                            actions={
-                                <Button variant='outline' onClick={() => navigate('/settings/analytics', {crossApp: true})}>
-                                    Open settings
-                                </Button>
-                            }
-                            description='Enable member source tracking in settings to see which content drives member growth.'
-                            title='Member sources have been disabled'
-                        >
-                            <LucideIcon.Activity />
-                        </EmptyIndicator>
+                        <DisabledSourcesIndicator />
                     </TableCell>
                 </TableRow>
             </TableBody>
